requirements.txt
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
Deprecated==1.2.13
Flask==2.0.2
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.3
jwcrypto==1.0
MarkupSafe==2.0.1
mysql-connector==2.2.9
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
python-jwt==3.3.1
six==1.16.0
Werkzeug==2.0.2
wrapt==1.13.3
